tweaked fxh6 sparks; reverted sysinfo class for a later impl
