+++ include/index_html.h
@@ -124,7 +124,7 @@ const char index_html[] PROGMEM = R"~~~(
     </main>
     <footer>
         <div>Version <span id="buildVersion"></span> built at <span id="buildTime"></span> on <span id="buildBranch"></span></div>
-        <div>Copyright &copy; 2023 Dan Luca. All rights reserved.</div>
+        <div>Copyright &copy; 2023<span id="curYear"></span> Dan Luca. All rights reserved.</div>
     </footer>
 </body>
