+++ src/www/index.html
@@ -123,7 +123,7 @@
     </main>
     <footer>
         <div>Version <span id="buildVersion"></span> built at <span id="buildTime"></span> on <span id="buildBranch"></span></div>
-        <div>Copyright &copy; 2023 Dan Luca. All rights reserved.</div>
+        <div>Copyright &copy; 2023<span id="curYear"></span> Dan Luca. All rights reserved.</div>
     </footer>
 </body>
