+++ src/www/pixel.js
@@ -19,6 +19,7 @@ let histOptions = {
 };
 
 $(() => {
+    $('#curYear').html(`-${new Date().getFullYear()}`);
 
     $.getJSON( "config.json")
     .done(function( data ) {
